Opened Spark session should be closed on exit
Hello,

In case of resources optimization, I thought we should close sparkR session as soon as we have finished working with data.
I've added a stop() call to the end of the code snippet, what do you think about it? Is it necessary? 

What happens when we shut down RStudio, does the session automatically wiped?